% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{copia}
\alias{copia}
\title{Title}
\usage{
copia(x)
}
\arguments{
\item{param1}{description}

\item{param1}{description}
}
\description{
Description
Description
Description
Description
}
\details{
Detais
Detais
Detais
}

