% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{f}
\alias{f}
\title{Title}
\usage{
f(x, type = c("ORIGINAL", "\%", "k", "M", "B"))
}
\arguments{
\item{param1}{description}
}
\description{
Description
Description
Description
Description
}
\details{
Detais
Detais
Detais
}
