% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\name{filter_duplicated_rows}
\alias{filter_duplicated_rows}
\title{Filter all rows that are not uniquely identified by a set of columns}
\usage{
filter_duplicated_rows(dt, cols)
}
\arguments{
\item{dt}{data frame}

\item{cols}{character vector with column names}
}
\description{
Description
Description
Description
Description
}
\details{
lorem ipsum
}
