% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificacoes.R
\docType{data}
\name{GIFP}
\alias{GIFP}
\title{Classificações GIFP}
\format{
An object of class \code{character} of length 4.
}
\usage{
GIFP
}
\description{
Vetor caracter com as seguintes classificações orçamentárias:
\itemize{
  \item Grupo de Despesa (GRUPO_COD)
  \item Identificador de Programa/Ação Governamental (IPG_COD)
  \item Fonte de Recurso (FONTE_COD)
  \item Identificador de Procedência e Uso (IPU_COD)
}
}
\keyword{datasets}
