% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkgs.R
\name{install_relatorios}
\alias{install_relatorios}
\title{Title}
\usage{
install_relatorios(ref = "master", password)
}
\arguments{
\item{param1}{description}
}
\description{
Description
Description
Description
Description
}
\details{
Detais
Detais
Detais
}
