% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\name{is_candidate_key}
\alias{is_candidate_key}
\title{Determine if a set of columns constitute a candidate key for a dataset}
\usage{
is_candidate_key(dt, cols)
}
\arguments{
\item{dt}{data frame}

\item{cols}{character vector with column names}
}
\description{
Description
Description
Description
Description
}
\details{
This function
}
