% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{Mescla de múltiplas bases}
\usage{
join(
  ...,
  by = NULL,
  value.var = "^vlr_",
  regex = TRUE,
  bind = c("rows", "columns"),
  idcol = NULL
)
}
\arguments{
\item{...}{bases separadas por virgulas que deverão ser mescladas}

\item{by}{vetor com as colunas presentes em todas as bases que serão utilizadas
para mescla. Se \code{NULL} serão utilizadas as colunas comuns que não forem
utilizadas como \code{value.var}}

\item{value.var}{variáveis de valor que serão agregadas pela função \code{sum}}

\item{regex}{\code{value.var} deve ser interpretado como uma expressão regular?}

\item{bind}{argumento que indica se as bases deverão ser empilhadas (rows) ou não (columns)}

\item{idcol}{argumento repassado a função \code{\link{rbindlist}}}
}
\description{
Função para realizar mescla de bases que sigam as convenções de nome das
colunas da AID
}
\details{
A coluna .id somente pode ser adicionada a base de dados final
 caso as bases sejam empilhadas (bind = "rows").
 Além disso, caso as bases sejam informadas com nome (ie. join(BASE = dados)),
 essa informação será utilizada para identificação das bases na coluna .id.
}
\examples{
receita <- data.frame(ano = c(2014, 2015), vlr_rec = c(100, 250))
despesa <- data.frame(ano = c(2014, 2015), vlr_desp = c(50, 25))

join(receita, despesa)
##     ano vlr_rec vlr_desp
## 1: 2014    100       0
## 2: 2015    250       0
## 3: 2014      0      50
## 4: 2015      0      25

join(receita, despesa, idcol = TRUE)
##    .id  ano vlr_rec vlr_desp
## 1:   1 2014    100       0
## 2:   1 2015    250       0
## 3:   2 2014      0      50
## 4:   2 2015      0      25

join(REC = receita, DESP = despesa, idcol = TRUE)
##     .id  ano vlr_rec vlr_desp
## 1:  REC 2014    100       0
## 2:  REC 2015    250       0
## 3: DESP 2014      0      50
## 4: DESP 2015      0      25

join(REC = receita, DESP = despesa, idcol = "base")
##    base  ano vlr_rec vlr_desp
## 1:  REC 2014    100       0
## 2:  REC 2015    250       0
## 3: DESP 2014      0      50
## 4: DESP 2015      0      25

#--------------------------------------------------------------------
exec <- data.frame(ano = c(2014, 2015), vlr_liq = c(100, 250))
rp <- data.frame(ano = c(2014, 2015), vlr_liq_rp = c(50, 25))

join(exec, rp, value.var = "vlr_liq")
##     ano vlr_liq vlr_liq_rp
## 1: 2014    100         0
## 2: 2015    250         0
## 3: 2014      0        50
## 4: 2015      0        25

join(exec, rp, value.var = "vlr_liq", regex = FALSE)
##     ANO VL_LIQ
## 1: 2014    100
## 2: 2015    250
## 3: 2014      0
## 4: 2015      0
}
