% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_base.R
\name{load_base}
\alias{load_base}
\title{Carrega e mescla bases de dados da execução orçamentária}
\usage{
load_base(base, dir = "~/../.execucao/")
}
\arguments{
\item{base}{identificador da base de dados (ie. exec_rec)}

\item{dir}{diretório que armazena as bases}
}
\description{
Carrega e mescla bases de dados da execução orçamentária
}
