% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index.R
\name{make_date}
\alias{make_date}
\title{Cria datas (YYYY-MM-DD) a partir de anos e meses}
\usage{
make_date(x, y)
}
\arguments{
\item{x}{coluna ano no formato YYYY}

\item{y}{coluna mes no formato MM}
}
\description{
Description
Description
Description
Description
}
\details{
O dia inserido será sempre 1
}

