% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index.R
\name{make_index}
\alias{make_index}
\title{Title}
\usage{
make_index(start, h, x, end, class = c("yearmon", "Date"))
}
\arguments{
\item{param1}{description}

\item{param1}{description}
}
\description{
Description
Description
Description
Description
}
\details{
Detais
Detais
Detais
}

