% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{p}
\alias{p}
\title{Title}
\usage{
p(x, type = c("ORIGINAL", "\%", "k", "M", "B"))
}
\arguments{
\item{param1}{description}

\item{param1}{description}
}
\description{
Description
Description
Description
Description
}
\details{
Detais
Detais
Detais
}

