% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_table.R
\name{prop_table}
\alias{prop_table}
\title{Title}
\usage{
prop_table(data, formula)
}
\arguments{
\item{param1}{description}

\item{param1}{description}
}
\description{
Description
Description
Description
Description
}
\details{
Detais
Detais
Detais
}
