% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxas_crescimento.R
\name{tx_cresc}
\alias{tx_cresc}
\title{Title}
\usage{
tx_cresc(x)
}
\arguments{
\item{param1}{description}
}
\description{
Description
Description
Description
Description
}
\details{
Detais
Detais
Detais
}
