% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index.R
\name{make_index}
\alias{make_index}
\title{Cria vetor de datas}
\usage{
make_index(start, h, x, end, class = c("date", "yearmon", "character"))
}
\arguments{
\item{start}{data de inicio no formato YYYY-MM}

\item{h}{numero para determinar tamanho do vetor}

\item{x}{objeto para determinar tamanho do vetor}

\item{end}{data final no formato YYYY-MM}

\item{class}{classe de retorno do vetor}
}
\description{
Description
Description
Description
Description
}
\details{
Somente h, x ou end devem ser especificados
}
